backend/requirements.txt
django
djangorestframework
django-environ
psycopg2-binary
django-cors-headers
Pillow
django_filters
drf-spectacular
